Add region function
