+++ oci/table_oci_core_nat_gateway.go
@@ -232,11 +232,3 @@ func natGatewayTags(_ context.Context, d *transform.TransformData) (interface{},
 
 	return tags, nil
 }
-
-// Extract OCI region name from the resource id
-func ociRegionName(_ context.Context, d *transform.TransformData) (interface{}, error) {
-	id := types.SafeString(d.Value)
-	splittedId := strings.Split(id, ".")
-	regionName := oci_common.StringToRegion(types.SafeString(splittedId[3]))
-	return regionName, nil
-}
