+++ oci/utils.go
@@ -4,12 +4,15 @@ import (
 	"context"
 	"math"
 	"strconv"
+	"strings"
 	"time"
 
 	oci_common "github.com/oracle/oci-go-sdk/v36/common"
 	"github.com/oracle/oci-go-sdk/v36/objectstorage"
 	"github.com/turbot/go-kit/helpers"
+	"github.com/turbot/go-kit/types"
 	"github.com/turbot/steampipe-plugin-sdk/plugin"
+	"github.com/turbot/steampipe-plugin-sdk/plugin/transform"
 )
 
 type nameSpace struct {
@@ -77,4 +80,12 @@ func getExponentialBackoffRetryPolicy(n uint, fn func(r oci_common.OCIOperationR
 	}
 	policy := oci_common.NewRetryPolicy(n, fn, exponentialBackoff)
 	return &policy
+}
+
+// Extract OCI region name from the resource id
+func ociRegionName(_ context.Context, d *transform.TransformData) (interface{}, error) {
+	id := types.SafeString(d.Value)
+	splittedId := strings.Split(id, ".")
+	regionName := oci_common.StringToRegion(types.SafeString(splittedId[3]))
+	return regionName, nil
 }
